Refracted Sex class with minor modifications
